+++ server/app/classes/game/solo-game.ts
@@ -130,15 +130,22 @@ export class SoloGame extends Game {
     delay = () => new Promise(res => setTimeout(res, 500));
 
     private endGame(): void {
-        this.endDate = new Date().getTime();
-        clearInterval(this.gameTimerInterval);
-        this.guessesLeft = 0;
-        this.vPlayer.stopDrawing();
-        this.socketService.getSocket().to(this.id).emit('endGame', { "finalScore": this.score, "virtualPlayerDrawings": this.pastVirtualDrawings, "virtualPlayerIds": this.pastVirtualDrawingsId });
-        this.vPlayer.sayEndSoloGame(this.score);
-        this.socketService.getSocket().to(this.id).emit('message', { "user": { username: "System", avatar: -1 }, "text": "La partie est maintenant terminée!", "timestamp": 0, "textColor": "#2065d4", chatId: this.id });
-        this.statsService.updateStats(this.gameName, this.gameType, this.getPlayers(), [this.score], this.startDate, this.endDate);
-        this.chatManagerService.deleteChat(this.id);
+        if(this.isGameEnded){
+            return;
+        }
+        else{
+            this.isGameEnded = true;
+            this.gameEnded.next(true);
+            this.endDate = new Date().getTime();
+            clearInterval(this.gameTimerInterval);
+            this.guessesLeft = 0;
+            this.vPlayer.stopDrawing();
+            this.socketService.getSocket().to(this.id).emit('endGame', { "finalScore": this.score, "virtualPlayerDrawings": this.pastVirtualDrawings, "virtualPlayerIds": this.pastVirtualDrawingsId });
+            this.vPlayer.sayEndSoloGame(this.score);
+            this.socketService.getSocket().to(this.id).emit('message', { "user": { username: "System", avatar: -1 }, "text": "La partie est maintenant terminée!", "timestamp": 0, "textColor": "#2065d4", chatId: this.id });
+            this.statsService.updateStats(this.gameName, this.gameType, this.getPlayers(), [this.score], this.startDate, this.endDate);
+            this.chatManagerService.deleteChat(this.id);
+        }
     }
 
     getPlayers(): any {
